Report JavaScript Parsing Errors
This addresses some NullPointerExceptions arriving.

The general philosophy is as follows:
JS: When parsing the entire JS, fail the run on a JS parse error.

String JS/Inline JS: Notify, but do a pass-through of the code without renaming.
This allows consumers to use expressions the renamer doesn't yet recognize.