Support JavaScript 'use strict' Directive
The Closure Compiler treats the 'use strict' directive as a property of a node.
CodeBuilder doesn't consider directives during its code generation. Instead, it
inserts the 'use strict' directive depending on its language mode.

The fix checks for a 'use strict' directive and requests ECMASCRIPT6_STRICT
if it is indeed set.